docs(TOPATCH)!: Major +comments, minor +refactor
(!) Also added a new Hathora.Demos.Shared asmdef
* To be patched @ hathora-unity.feat/mirror
* Namespace fixes
* Base fixes
