+++ Assets/Hathora/Core/Scripts/Editor/Common/HathoraEditorUtils.cs
@@ -15,7 +15,13 @@ using Debug = UnityEngine.Debug;
 namespace Hathora.Core.Scripts.Editor.Common
 {
     /// <summary>
-    /// Editor script to add the Hathora banner to an editor window
+    /// Editor script to add useful Editor utils for Hathora:
+    /// - Color and hyperlink constants.
+    /// - Hathora banner to an editor window.
+    /// - Common style utils.
+    /// - File utils, including ".hathora" dir helpers. 
+    /// - Terminal Process cmd handling.
+    /// - DateTime parsing.
     /// </summary>
     public static class HathoraEditorUtils
     {
