+++ Assets/Hathora/Core/Scripts/Editor/Server/HathoraServerConfigFinder.cs
@@ -9,6 +9,9 @@ using UnityEngine;
 
 namespace Hathora.Core.Scripts.Editor.Server
 {
+    /// <summary>
+    /// Handles the ConfigFinder window @ top menu `Hathora/ConfigFinder`. 
+    /// </summary>
     [InitializeOnLoad]
     public class HathoraServerConfigFinder : EditorWindow
     {
