+++ Assets/Hathora/Core/Scripts/Runtime/Common/Extensions/UnityWebRequestAsyncOperationExtensions.cs
@@ -6,10 +6,15 @@ using UnityEngine.Networking;
 namespace Hathora.Core.Scripts.Runtime.Common.Extensions
 {
     /// <summary>
-    /// Allow an awaitable Task
+    /// Extension methods for UnityWebRequests.
     /// </summary>
     public static class UnityWebRequestAsyncOperationExtensions
     {
+        /// <summary>
+        /// Allow an awaitable Task. Normally, this is a coroutine.
+        /// </summary>
+        /// <param name="asyncOperation"></param>
+        /// <returns></returns>
         public static Task<UnityWebRequest> AsTask(this UnityWebRequestAsyncOperation asyncOperation)
         {
             var completionSource = new TaskCompletionSource<UnityWebRequest>();
