+++ Assets/Hathora/Core/Scripts/Runtime/Common/Models/ApplicationAuthConfigurationWrapper.cs
@@ -6,6 +6,8 @@ using Hathora.Cloud.Sdk.Model;
 namespace Hathora.Core.Scripts.Runtime.Common.Models
 {
     /// <summary>
+    /// Hathora SDK model wrapper to add [Serializable] support.
+    /// 
     /// This is a wrapper for Hathora SDK's `ApplicationWithDeployment` model.
     /// We'll eventually replace this with a [Serializable] revamp of the model.
     /// </summary>
