+++ Assets/Hathora/Core/Scripts/Runtime/Common/Models/ConnectionInfoV2Wrapper.cs
@@ -10,6 +10,8 @@ using UnityEngine.Serialization;
 namespace Hathora.Core.Scripts.Runtime.Common.Models
 {
     /// <summary>
+    /// Hathora SDK model wrapper to add [Serializable] support.
+    /// 
     /// This is a wrapper for Hathora SDK's `ConnectionInfoV2` model.
     /// We'll eventually replace this with a [Serializable] revamp of the model.
     /// </summary>
