+++ Assets/Hathora/Core/Scripts/Runtime/Common/Models/HathoraEnvVars.cs.meta
