+++ Assets/Hathora/Core/Scripts/Runtime/Server/ApiWrapper/HathoraServerDeployApi.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using Hathora.Cloud.Sdk.Api;
 using Hathora.Cloud.Sdk.Client;
 using Hathora.Cloud.Sdk.Model;
+using Hathora.Core.Scripts.Runtime.Common.Models;
 using Hathora.Core.Scripts.Runtime.Common.Utils;
 using Hathora.Core.Scripts.Runtime.Server.Models;
 using UnityEngine;
