+++ Assets/Hathora/Demos/Shared/Hathora.Demos.Shared.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 49fa1b98c11fb0a4ab3f6ec21a90865a
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
